Remove 5gk and 5gj mappings
This is a remnant from the original colemak.vim that I didn't remove;
I haven't used these mappings for a long time now. Besides, there are
better ways to navigate code.

Additionally, remove a comment that was both a duplicate comment and
completely unrelated to the code it was right above.