Remove custom PageUp/PageDown mappings
It seems like the idea Shai Coleman (author of the original
`colemak.vim`) had was that PageUp/PageDown had the problem of not going
all the way up to the top of the document (and vice versa for PageDown).

This behaviour is a difference from other editors, and I recall that
there have been many things in the original `colemak.vim` that aimed to
"rectify" such differences (e.g. swapping Normal mode `p` with `P`).

I'm currently used to the behaviour of going all the way to the
top/bottom, but I'm strongly leaning to cutting this in favor of just
using the built-in <PageUp> and <PageDown> functionality, to cut down on
the custom code.