Remove undo mapping Visual and Select
I'm guessing this was a case of Shai Coleman copy-pasting according to
the scheme of having every mapping apply both in Normal and in other
modes, but here I don't think it makes any sense.

I noticed that this interferes with some plugin mappings that I wanted
to set, and I see no reason for keeping these.