Replace `yy`/`cc` mappings with added onoremap mappings
The idea of `yy` and `cc` is to operate on a whole line. However, the
same result can be achieved by simply have the o-mapping for the same
key to be added. It seems to me like this is the more uniform and
correct way of achieving the same result.